Change commands' descriptions and some terminology to be more consistent
This commit changes commands' descriptions to be more consistent with Clap's new style (e.g., "Prints" -> "Print") and more descriptive.

It also changes a bunch of strings and documentation to be a bit more consistent in terminology (site / web app / PWA). "Site/website" now represents a website available on the web, which you can install to your computer as a "web app" (PWA). Note that "site" is still used a bit inconsistently inside the code, and in commands and connector requests to avoid breaking backward compatibility.