Use "double option" pattern for parameters when updating sites/profiles
Both the CLI and the API now properly detect when a parameter was specified. If the parameter is skipped, its value remains unchanged. If it is set to `None`, it uses default/manifest value. If it is set to any other value, that value is stored.

This also removes internal `store_none_values` parameter as it is not needed anymore.