Implement Schema.org and add web app manifest
This commit implements JSON-LD schema for the homepage, providing the general information about the website and the project, and JSON-LD breadcrumbs for navigation. Additionally, it transforms the FAQ page into a valid microdata format.

This commit also adds a basic web app manifest for the documentation website. More icons will be added later.