Fix compatibility with Firefox 134 (fixes #600)
Because of BZ-1930654 / PHAB-228667, `window._gBrowser` is no longer available immediately on startup, which breaks some code that relies on it. To fix the issue, we need to delay the affected code after the browser has been initialized.

This is quite a hacky solution. In the future, we should rewrite how loading works and find a better solution.