Specify browsing activity transmission for the extension
According to the Firefox Add-ons Policies, communication with the native messaging host is considered data collection and transmission. To comply with that, the extension manifest now declares the transmission of the browsing activity.

No changes regarding the actual transmission have been made. No personal data leave the computer or are sent to third parties.